refactor: separate a misconfig scanner
